+++ internal/util-cache/src/main/scala/sbt/util/internal/SeparatedCache.scala
@@ -1,7 +1,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2009 Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 import Types.:+:
 import sbinary.{ DefaultProtocol, Format, Input, Output => Out }
