+++ internal/util-logging/src/main/scala/sbt/util/internal/ConsoleOut.scala
@@ -1,4 +1,4 @@
-package sbt
+package sbt.util.internal
 
 import java.io.{ BufferedWriter, PrintStream, PrintWriter }
