+++ internal/util-logging/src/main/scala/sbt/util/internal/GlobalLogging.scala
@@ -1,7 +1,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2010 Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 import java.io.{ File, PrintWriter }
