+++ internal/util-logging/src/main/scala/sbt/util/internal/Level.scala
@@ -1,7 +1,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2008, 2009 Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 /**
  * An enumeration defining the levels available for logging.  A level includes all of the levels
