+++ internal/util-logging/src/main/scala/sbt/util/internal/LogEvent.scala
@@ -1,7 +1,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2008, 2009 Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 sealed trait LogEvent extends NotNull
 final class Success(val msg: String) extends LogEvent
