+++ internal/util-logging/src/main/scala/sbt/util/internal/MainLogging.scala
@@ -1,4 +1,4 @@
-package sbt
+package sbt.util.internal
 
 import java.io.PrintWriter
