+++ internal/util-logging/src/main/scala/sbt/util/internal/MultiLogger.scala
@@ -2,7 +2,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2008, 2009, 2010 Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 // note that setting the logging level on this logger has no effect on its behavior, only
 //   on the behavior of the delegates.
