+++ internal/util-logging/src/test/scala/Escapes.scala
@@ -1,4 +1,4 @@
-package sbt
+package sbt.util.internal
 
 import org.scalacheck._
 import Prop._
